pyproject.toml
[project]
name = "unif"
version = "0.1.0"
description = "University LMS Notification"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.18.0",
    "aiohttp>=3.11.13",
    "beautifulsoup4>=4.13.3",
    "django-celery-beat>=2.7.0",
    "django>=5.1.7",
    "psycopg[binary,pool]>=3.2.5",
    "redis>=5.2.1",
    "flower>=2.0.1",
    "python-dotenv>=1.0.1",
    "whitenoise>=6.9.0",
    "gunicorn>=23.0.0",
    "celery>=5.4.0",
]

[dependency-groups]
dev = [
    "mkdocs-material>=9.6.7",
    "mkdocs>=1.6.1",
    "mkdocstrings[python]>=0.28.2",
    "ruff>=0.9.9",
    "mkdocs-include-markdown-plugin>=7.1.5",
]

[tool.ruff]
line-length = 79
